requirements.txt
pytorch >= 1.6.0
torchvision >= 0.7.0
tqdm >= 4.55.0
pillow >= 8.3.1
